New login slots and auth plugins
