+++ client/coral-sign-in/containers/LoginContainer.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import Slot from 'coral-framework/components/Slot';
+
+export const LoginContainer = () => (
+  <Slot fill="login"/>
+)
