+++ plugins/coral-plugin-auth/client/index.js
@@ -1,7 +1,9 @@
 import SignInButton from './components/SignInButton';
+import SignInContainer from './components/SignInContainer';
 
 export default {
   slots: {
-    stream: [SignInButton]
+    stream: [SignInButton],
+    login: [SignInContainer]
   }
 };
